first task 0.2

+++ modules/task_1/Kosterin_A_IntegMonteMPI/integMonteMPI.h
@@ -3,9 +3,7 @@
 #ifndef MODULES_TASK_1_KOSTERIN_A_INTEGMONTEMPI_INTEGMONTEMPI_H_
 #define MODULES_TASK_1_KOSTERIN_A_INTEGMONTEMPI_INTEGMONTEMPI_H_
 
-#include <iostream>
-#include <stdio.h>
-#include <Windows.h>
+#include <random>
 
-double monteCarlo(double low, double high, int count, double (*f)(double));
+double monteCarlo(int low, int high, int count, double (*f)(double));
 #endif // MODULES_TASK_1_KOSTERIN_A_INTEGMONTEMPI_INTEGMONTEMPI_H_
